# new diff/commit search: improve performance for case-insensitive search