--- a/internal/gitserver/search/match_tree.go
+++ b/internal/gitserver/search/match_tree.go
@@ -2,6 +2,7 @@ package search
 
 import (
 	"bytes"
+	"unicode/utf8"
 
 	"github.com/cockroachdb/errors"