--- a/internal/search/casetransform/lower_amd64.go
+++ b/internal/search/casetransform/lower_amd64.go
@@ -1,0 +1,4 @@
+package casetransform
+
+// implemented in assembly, see lower_amd64.s
+func BytesToLowerASCII(dst, src []byte)