--- a/internal/search/casetransform/lower_other.go
+++ b/internal/search/casetransform/lower_other.go
@@ -1,0 +1,6 @@
+//go:build !amd64
+// +build !amd64
+
+package casetransform
+
+var BytesToLowerASCII = bytesToLowerASCIIgeneric