--- a/internal/search/casetransform/regexp.go
+++ b/internal/search/casetransform/regexp.go
@@ -5,6 +5,18 @@ import (
 	"regexp/syntax"
 )
 
+// Regexp is a light wrapper over *regexp.Regexp that optimizes for case-insensitive search.
+//
+// Case-insensitive search using *regexp.Regexp and `(?i)` meta tags is quite
+// slow. To mitigate the performance cost of case-insensitive search, we
+// transform regexp patterns to their lower-case equivalent (LowerRegexpASCII),
+// and transform the search content to its lower-case equivalent (BytesToLowerASCII)
+// before matching the pattern to the content.
+//
+// This type encodes the requirements that, if ignoreCase is set:
+// 1) The regexp pattern is transformed into its lower-case equivalent
+// 2) The content to be searched is transformed into its lower-case equivalent
+// 3) A re-usable buffer is passed in to the match methods to encourage buffer re-use
 type Regexp struct {
 	re         *regexp.Regexp
 	ignoreCase bool