… cmd/searcher/search/search_regex_test.go
… internal/gitserver/search/search_test.go
… internal/gitserver/search/testdata/large_diff.txt
… internal/gitserver/search/testdata/small_diff.txt
… internal/search/casetransform/lower_test.go