@camdencheek 2021-09-27
> Is my read of the benchmarks correct here

Yes! Now, to be fair, this is partly re-claiming performance regression from moving the regex evaluation completely into go. Git's regex implementation is at least a little more performant than go's for case insensitive, but it still struggles. I haven't done any serious benchmarking comparison yet (still chasing lower-hanging fruit), but based on anecdotal data, this is a bit faster than using git's case-insensitive flag `-i`.

Also, this only does lower-casing for ASCII characters, so we trade a little correctness for a pretty significant amount of performance (same as searcher).